feat: memo 라우터 작성
update와 delete는 쿼리의 조건에 accountId를 비교하는 방식으로 구현

그 외 지난 커밋에서 반영안된 파일들

+++ src/services/MemoService.ts
@@ -1,4 +1,4 @@
-import { DataAccessError } from 'errors/DataAccessError';
+import DataAccessError from 'errors/DataAccessError';
 import Memo from 'models/Memo.model';
 import { Service } from 'typedi';
 
@@ -12,15 +12,15 @@ export default class MemoService {
 		return Memo.findAll({ offset, limit: this.limit });
 	}
 
-	public async insert(userid: number, body: string): Promise<Memo> {
+	public async insert(accountId: number, body: string): Promise<Memo> {
 		try {
 			const head = body.substring(0, 20);
-			const result = await Memo.create({ userid, head, body }, { raw: true });
+			const result = await Memo.create({ accountId, head, body }, { raw: true });
 			if (!result) throw new DataAccessError('Memo does not saved');
 			return result;
 		} catch (error) {
 			if (typeof error == 'string' || !(error instanceof Error)) throw error;
-			if (error.name === 'ForeignKeyConstraintError') throw new DataAccessError(`${userid} has some problem`);
+			if (error.name === 'ForeignKeyConstraintError') throw new DataAccessError(`${accountId} has some problem`);
 			throw error;
 		}
 	}
@@ -32,15 +32,15 @@ export default class MemoService {
 		//todo fetch comments
 	}
 
-	public async updateBody(id: number, body: string): Promise<Memo> {
+	public async updateBody(id: number, accountId: number, body: string): Promise<Memo> {
 		const head = body.substring(0, 20);
-		const [affectedCount, affectedRow] = await Memo.update({ head, body }, { where: { id }, returning: true });
+		const [affectedCount, affectedRow] = await Memo.update({ head, body }, { where: { id, accountId }, returning: true });
 		if (affectedCount === 0) throw new DataAccessError('Memo does not exist');
 		return affectedRow[0];
 	}
 
-	public async delete(id: number): Promise<void> {
-		const result = await Memo.destroy({ where: { id } });
+	public async delete(id: number, accountId: number): Promise<void> {
+		const result = await Memo.destroy({ where: { id, accountId } });
 		if (result === 0) throw new DataAccessError('Memo does not exist');
 		//todo remove with comments
 	}
